+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Change Log
 
+### Changed
+- node v22.12 ; npm v10.9; rollup v2.79
+
 ## [1.70.0] 2026-04-04
 ### Added
 - `remember`: 'forget' a key